pyproject.toml
[project]
name = "ih"
version = "0.0.1"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
#readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies=["fastapi[standard] (>=0.120.4,<0.121.0)", "uvicorn (>=0.38.0,<0.39.0)", "alembic (>=1.17.1,<2.0.0)", "apscheduler (>=3.11.1,<4.0.0)", "pyjwt (>=2.10.1,<3.0.0)", "sqlmodel (>=0.0.27,<0.0.28)", "pydantic-to-typescript (>=2.0.0,<3.0.0)", "psycopg2-binary (>=2.9.11,<3.0.0)", "python-dotenv (>=1.2.1,<2.0.0)", "python-multipart (>=0.0.20,<0.0.21)", "pydantic-settings (>=2.11.0,<3.0.0)", "fastapi-utils (>=0.8.0,<0.9.0)", "pytest (>=8.4.2,<9.0.0)", "jinja2 (>=3.1.6,<4.0.0)", "typing-inspect (>=0.9.0,<0.10.0)", "pwdlib[argon2] (>=0.3.0,<0.4.0)", "sqlalchemy-utc (>=0.14.0,<0.15.0)", "authlib (>=1.6.5,<2.0.0)", "itsdangerous (>=2.2.0,<3.0.0)", "pillow (>=12.0.0,<13.0.0)", "pytest-cov (>=7.0.0,<8.0.0)", "requests (>=2.32.5,<3.0.0)", "beautifulsoup4 (>=4.14.3,<5.0.0)"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


# Console script for running the app
[tool.poetry.scripts]
ih = "ih.main:main"

